deriv_bot_fixed/requirements.txt
websockets>=12.0
aiohttp>=3.9.0
aiosqlite>=0.20.0
python-telegram-bot>=21.0
python-dotenv>=1.0.0
google-genai
numpy